feat!: use gradient method for the optimisation part (#31)
We change the optimisation algorithm (computation of crossed modalities probabilities) for a self implemented one, in order to get rid of the maxentropy package which is no longer maintained. Enrichment classes are given more consistent name.
The global quality of the resulting population is improved by this new implementation (better fit to the marginal distributions).

Former modules, classes and tests are kept in this commit to provide a comparison environment if needed. They will be removed later.

CAUTION : these changes introduced a problem where the test_bhepop2_enrich test would pass locally but not on github. See #34 

BREAKING CHANGE: 
- change in the optimisation algorithm means that the results will change, hence breaking reproducibility
- main enrichment class is renamed Bhepop2Enrichment

Close: #28